requirements.txt
airtest==1.0.26
BeautifulReport==0.1.2
bitfinex==0.2.6
cigam==0.0.3
Click==7.0
fire==0.1.3
Flask==0.12.2
Flask-Admin==1.5.0
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
greenlet==0.4.15
httplib2==0.9.2
ipython==7.8.0
pocoui==1.0.76
progress==1.5
prompt-toolkit==2.0.9
py==1.8.0
Pygments==2.4.2
PyMySQL==0.9.3
PyNaCl==1.3.0
PyQt5==5.13.0
PyQt5-sip==4.19.18
pytesseract==0.2.6
pyzmq==18.1.0
requests==2.22.0
rethinkdb==2.4.3.post1
selenium==3.141.0
SQLAlchemy==1.3.4


